+++ example/project/__test__/App.test.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import { render } from '@testing-library/react';
 import { chain, nullChain } from '@/chain';
+// eslint-disable-next-line import/extensions
 import App from '../app/App';
 
 describe('App', () => {