+++ packages/workflow/webpack.config.production.js
@@ -15,10 +15,10 @@ const html = require('./html');
 
 process.noDeprecation = true;
 
-const plugin = settings => {
+const plugin = (settings) => {
   const babelrcPath = path.join(settings.project(), '.babelrc');
   const babelrcExists = fs.existsSync(babelrcPath);
-  const resolveApp = relativePath => path.resolve(settings.app(), relativePath);
+  const resolveApp = (relativePath) => path.resolve(settings.app(), relativePath);
 
   function getVersion() {
     return settings.pkg().version || 'N/A';
@@ -27,11 +27,16 @@ const plugin = settings => {
   const config = {
     mode: 'production',
 
+    // Using an explicit browserslist query will override any project configs
+    // This is fine because we need to ensure IE 11 is being targeted
+    target: 'browserslist: defaults, ie 11',
+
     context: settings.app(),
 
     entry: {
       index: [
         require.resolve('react-app-polyfill/ie11'),
+        require.resolve('polyfill-array-includes'),
         require.resolve('navigator.sendbeacon'),
         resolveModule(resolveApp, 'index')
       ]
@@ -40,37 +45,73 @@ const plugin = settings => {
     optimization: {
       splitChunks: {
         cacheGroups: {
+          // https://webpack.js.org/plugins/split-chunks-plugin/#optimizationsplitchunks
+          // Add defaultVendors and default so we don't override webpack 5 defaults
+          // When files paths are processed by webpack, they always contain / on Unix systems and \ on Windows.
+          // That's why using [\\/] in {cacheGroup}.test fields is necessary to represent a path separator.
+          // / or \ in { cacheGroup }.test will cause issues when used cross- platform.
+          defaultVendors: {
+            test: /[/\\]node_modules[/\\]/,
+            idHint: 'defaultVendors',
+            chunks: 'all',
+            priority: -10
+          },
+          default: {
+            minChunks: 2,
+            priority: -20,
+            reuseExistingChunk: true
+          },
+          // Create a chunk for react and react-dom since they shouldn't change often
+          // https://webpack.js.org/guides/caching/#extracting-boilerplate
+          // https://webpack.js.org/plugins/split-chunks-plugin/#split-chunks-example-3
+          react: {
+            test: /[/\\]node_modules[/\\](react|react-dom)[/\\]/,
+            idHint: 'vendors',
+            chunks: 'all',
+            priority: 10
+          },
+          // Create a chunk for lodash or any of its separate packages
+          lodash: {
+            // Should capture lodash/, lodash-es/, lodash.whateverModule/
+            test: /[/\\]node_modules[/\\](lodash([.-])?\w*?)[/\\]/,
+            idHint: 'vendors',
+            chunks: 'all',
+            priority: 1
+          },
+          moment: {
+            test: /[/\\]node_modules[/\\](moment)[/\\]/,
+            idHint: 'vendors',
+            chunks: 'all',
+            priority: 2
+          },
           styles: {
-            name: 'styles',
+            idHint: 'styles',
             test: /\.css$/,
             chunks: 'all',
             enforce: true
-          },
-          commons: {
-            chunks: 'initial',
-            minChunks: 2
-          },
-          vendor: {
-            test: /node_modules/,
-            chunks: 'initial',
-            name: 'vendor',
-            priority: 10,
-            enforce: true
           }
+          // TODO: Add cacheGroup for Availity packages in node_modules ?
         }
       },
-      minimizer: []
+      minimizer: [],
+
+      // To extract boilerplate like runtime and manifest info
+      // Should aid caching by keeping filenames consistent if content doesn't change
+      runtimeChunk: 'single',
+
+      // Keeps vendor hashes consistent between builds where local dependency imports change the order of resolution
+      // https://webpack.js.org/guides/caching/#module-identifiers
+      // https://webpack.js.org/configuration/optimization/#optimizationmoduleids
+      moduleIds: 'deterministic'
     },
 
     output: {
       path: settings.output(),
       filename: settings.fileName(),
       chunkFilename: settings.chunkFileName(),
-      // TODO: remove this when upgrading to webpack 5
-      futureEmitAssets: true,
-      devtoolModuleFilenameTemplate: info =>
+      devtoolModuleFilenameTemplate: (info) =>
         `webpack:///${path.relative(settings.project(), info.absoluteResourcePath)}${
-        info.loaders ? `?${info.loaders}` : ''
+          info.loaders ? `?${info.loaders}` : ''
         }`
     },
 
@@ -85,7 +126,10 @@ const plugin = settings => {
         path.join(__dirname, 'node_modules')
       ],
       symlinks: true,
-      extensions: ['.js', '.jsx', '.ts', '.tsx', '.json', '.css', 'scss']
+      extensions: ['.js', '.jsx', '.ts', '.tsx', '.json', '.css', 'scss'],
+      fallback: {
+        path: require.resolve('path-browserify')
+      }
     },
 
     // This set of options is identical to the resolve property set above,
@@ -97,24 +141,90 @@ const plugin = settings => {
 
     module: {
       rules: [
+        // "oneOf" will traverse all the following loaders until it finds a match.
+        // If no loader matches it will use the default file loader at the end of this list.
         {
-          test: /\.(js|mjs|jsx|ts|tsx)$/,
-          include: settings.include(),
-          use: [
+          oneOf: [
+            // solution to process.cwd() is undefined in @availity/spaces -> react-markdown -> vfile
+            // https://github.com/remarkjs/react-markdown/issues/339#issuecomment-683199835
+            // Needed for @availity/spaces compatibility with Webpack 5
             {
+              test: /[/\\]node_modules[/\\]vfile[/\\]core\.js/,
+              use: [
+                {
+                  loader: 'imports-loader',
+                  options: {
+                    type: 'commonjs',
+                    imports: ['single process/browser process']
+                  }
+                }
+              ]
+            },
+            // Process application JS and user-specified paths with Babel.
+            {
+              test: /\.(js|mjs|jsx|ts|tsx)$/,
+              include: settings.include(),
+              use: [
+                {
+                  loader: 'babel-loader',
+                  options: {
+                    presets: [babelPreset],
+                    // This is a feature of `babel-loader` for webpack (not Babel itself).
+                    // It enables caching results in ./node_modules/.cache/babel-loader/
+                    // directory for faster rebuilds.
+                    cacheDirectory: true,
+                    // See https://github.com/facebook/create-react-app/issues/6846 for context on why cacheCompression is disabled
+                    cacheCompression: false,
+                    babelrc: babelrcExists
+                  }
+                }
+              ]
+            },
+            // Process any JS outside of the app with Babel.
+            // Unlike the application JS, we only compile the standard ES features.
+            {
+              test: /\.(js|mjs)$/,
+              exclude: [...settings.include(), /@babel(?:\/|\\{1,2})runtime/],
               loader: 'babel-loader',
               options: {
-                presets: [babelPreset],
-                cacheDirectory: settings.isDevelopment(),
-                babelrc: babelrcExists
+                babelrc: false,
+                configFile: false,
+                compact: false,
+                presets: [[require.resolve('babel-preset-react-app/dependencies'), { helpers: true }]],
+                cacheDirectory: true,
+                // See https://github.com/facebook/create-react-app/issues/6846 for context on why cacheCompression is disabled
+                cacheCompression: false
+              }
+            },
+            // Allows .mjs and .js files from packages of type "module" to be required without the extension
+            {
+              test: /\.m?js/,
+              resolve: {
+                fullySpecified: false
+              }
+            },
+            loaders.css.production,
+            loaders.scss.production,
+            loaders.fonts,
+            loaders.images,
+            // "file" loader makes sure those assets get served by WebpackDevServer.
+            // When you `import` an asset, you get its (virtual) filename.
+            // In production, they would get copied to the `build` folder.
+            // This loader doesn't use a "test" so it will catch all modules
+            // that fall through the other loaders.
+            {
+              loader: require.resolve('file-loader'),
+              // Exclude `js` files to keep "css" loader working as it injects
+              // its runtime that would otherwise be processed through "file" loader.
+              // Also exclude `html` and `json` extensions so they get processed
+              // by webpack's internal loaders.
+              exclude: [/\.(js|mjs|jsx|ts|tsx)$/, /\.html$/, /\.json$/],
+              options: {
+                name: 'static/media/[name].[contenthash:8].[ext]'
               }
             }
           ]
-        },
-        loaders.css.production,
-        loaders.scss.production,
-        loaders.fonts,
-        loaders.images
+        }
       ]
     },
     plugins: [
@@ -134,12 +244,12 @@ const plugin = settings => {
       new HtmlWebpackPlugin(html(settings)),
 
       // Ignore all the moment local files
-      new webpack.IgnorePlugin(/^\.\/locale$/, /moment$/),
+      new webpack.IgnorePlugin({ resourceRegExp: /^\.\/locale$/, contextRegExp: /moment$/ }),
 
       new CaseSensitivePathsPlugin(),
 
       new loaders.MiniCssExtractPlugin({
-        filename: 'css/[name]-[contenthash].css'
+        filename: 'css/[name]-[contenthash:8].chunk.css'
       }),
 
       new CopyWebpackPlugin({
@@ -160,8 +270,8 @@ const plugin = settings => {
       new TerserPlugin({
         terserOptions: {
           parse: {
-            // we want terser to parse ecma 8 code. However, we don't want it
-            // to apply any minfication steps that turns valid ecma 5 code
+            // We want terser to parse ecma 8 code. However, we don't want it
+            // to apply any minification steps that turns valid ecma 5 code
             // into invalid ecma 5 code. This is why the 'compress' and 'output'
             // sections only apply transformations that are ecma 5 safe
             // https://github.com/facebook/create-react-app/pull/4234
@@ -174,7 +284,12 @@ const plugin = settings => {
             // https://github.com/facebook/create-react-app/issues/2376
             // Pending further investigation:
             // https://github.com/mishoo/UglifyJS2/issues/2011
-            comparisons: false
+            comparisons: false,
+            // Disabled because of an issue with Terser breaking valid code:
+            // https://github.com/facebook/create-react-app/issues/5250
+            // Pending further investigation:
+            // https://github.com/terser-js/terser/issues/120
+            inline: 2
           },
           mangle: {
             safari10: true
@@ -189,10 +304,7 @@ const plugin = settings => {
         },
         // Use multi-process parallel running to improve the build speed
         // Default number of concurrent runs: os.cpus().length - 1
-        parallel: true,
-        // Enable file caching
-        cache: true,
-        sourceMap: true
+        parallel: true
       }),
       new OptimizeCSSAssetsPlugin({
         cssProcessorOptions: { zindex: false, reduceIdents: false }